refactor merge_attributes
