+++ tools/codegen/codespec/merge_attributes.go
@@ -2,6 +2,8 @@ package codespec
 
 import (
 	"sort"
+
+	"github.com/mongodb/terraform-provider-mongodbatlas/internal/common/autogen/stringcase"
 )
 
 // mergeNestedAttributes recursively merges nested attributes
@@ -177,4 +179,101 @@ func sortAttributes(attrs Attributes) {
 	sort.Slice(attrs, func(i, j int) bool {
 		return attrs[i].TFSchemaName < attrs[j].TFSchemaName
 	})
+}
+
+// mergeDataSourceAttributes merges path parameters with response attributes for data sources.
+// Path params are enforced as required; response attributes are marked computed (including all nested attributes).
+// Aliases are applied to both path params and response attributes during merge to properly detect duplicates.
+// If duplicates exist (same TFSchemaName after aliasing), Required always wins over Computed.
+func mergeDataSourceAttributes(pathParams, responseAttrs Attributes, aliases map[string]string) Attributes {
+	merged := make(map[string]*Attribute) // key by TFSchemaName
+
+	// Add path params as required (they identify the data source)
+	// Apply aliases to path params during merge
+	for i := range pathParams {
+		attr := pathParams[i] // create a copy
+		attr.ComputedOptionalRequired = Required
+		attr.ReqBodyUsage = OmitAlways
+
+		// Apply alias if configured
+		if alias, found := aliases[attr.APIName]; found {
+			attr.TFSchemaName = stringcase.ToSnakeCase(alias)
+			attr.TFModelName = stringcase.Capitalize(alias)
+		}
+
+		merged[attr.TFSchemaName] = &attr
+	}
+
+	// Add response attributes as computed (including all nested attributes)
+	// Apply aliases to response attributes during merge to detect duplicates with aliased path params
+	// If a duplicate exists and the existing one is Required, keep Required
+	for i := range responseAttrs {
+		attr := responseAttrs[i] // create a copy
+		setAttributeComputedRecursive(&attr)
+
+		// Apply alias if configured (same logic as path params)
+		if alias, found := aliases[attr.APIName]; found {
+			attr.TFSchemaName = stringcase.ToSnakeCase(alias)
+			attr.TFModelName = stringcase.Capitalize(alias)
+		}
+
+		if existing, found := merged[attr.TFSchemaName]; found {
+			// Duplicate found: keep Required over Computed (Required always wins)
+			if existing.ComputedOptionalRequired != Required {
+				merged[attr.TFSchemaName] = &attr
+			}
+			// else: existing is Required, keep it
+		} else {
+			merged[attr.TFSchemaName] = &attr
+		}
+	}
+
+	// Convert map to slice
+	result := make(Attributes, 0, len(merged))
+	for _, attr := range merged {
+		result = append(result, *attr)
+	}
+
+	sortAttributes(result)
+
+	return result
+}
+
+// setAttributeComputedRecursive sets an attribute and all its nested attributes to Computed.
+// This is used for data source attributes where all values come from the API response.
+func setAttributeComputedRecursive(attr *Attribute) {
+	attr.ComputedOptionalRequired = Computed
+	attr.ReqBodyUsage = OmitAlways
+
+	// Process nested attributes in ListNested
+	if attr.ListNested != nil {
+		for i := range attr.ListNested.NestedObject.Attributes {
+			setAttributeComputedRecursive(&attr.ListNested.NestedObject.Attributes[i])
+		}
+		sortAttributes(attr.ListNested.NestedObject.Attributes)
+	}
+
+	// Process nested attributes in SingleNested
+	if attr.SingleNested != nil {
+		for i := range attr.SingleNested.NestedObject.Attributes {
+			setAttributeComputedRecursive(&attr.SingleNested.NestedObject.Attributes[i])
+		}
+		sortAttributes(attr.SingleNested.NestedObject.Attributes)
+	}
+
+	// Process nested attributes in SetNested
+	if attr.SetNested != nil {
+		for i := range attr.SetNested.NestedObject.Attributes {
+			setAttributeComputedRecursive(&attr.SetNested.NestedObject.Attributes[i])
+		}
+		sortAttributes(attr.SetNested.NestedObject.Attributes)
+	}
+
+	// Process nested attributes in MapNested
+	if attr.MapNested != nil {
+		for i := range attr.MapNested.NestedObject.Attributes {
+			setAttributeComputedRecursive(&attr.MapNested.NestedObject.Attributes[i])
+		}
+		sortAttributes(attr.MapNested.NestedObject.Attributes)
+	}
 }
